+++ lib/plugins/aws/package/compile/events/iot.js
