+++ lib/plugins/index.js
@@ -44,7 +44,7 @@ module.exports = [
   require('./aws/package/compile/events/alb/index.js'),
   require('./aws/package/compile/events/alexaSkill.js'),
   require('./aws/package/compile/events/alexaSmartHome.js'),
-  require('./aws/package/compile/events/iot/index.js'),
+  require('./aws/package/compile/events/iot.js'),
   require('./aws/package/compile/events/cloudWatchEvent.js'),
   require('./aws/package/compile/events/cloudWatchLog.js'),
   require('./aws/package/compile/events/cognitoUserPool.js'),
